refactor: 폴더 구조 재배치 및 그에 따른 참조 경로 수정
