+++ src/domain/auth/mappers/mapLoginUserToUser.ts
@@ -1,5 +1,5 @@
 import type { User } from '@src/domain/user/Interface/User';
-import type { LoginUser } from '@src/pages/login/login.types';
+import type { LoginUser } from '@src/pages/login/types/login.types';
 
 export function mapLoginUserToUser(dto: LoginUser): User {
   return {
